+++ lib/ronin/support/binary/struct.rb
@@ -657,8 +657,9 @@ module Ronin
         #         :arm64, :arm64_le, :arm64_be] arch
         #   The desired architecture for the struct.
         #
-        # @param [:linux, :macos, :windows, :android,
-        #         :bsd, :freebsd, :openbsd, :netbsd] os
+        # @param [:linux, :macos, :windows,
+        #         :android, :apple_ios, :bsd,
+        #         :freebsd, :openbsd, :netbsd] os
         #   The Operating System (OS) for the struct.
         #
         # @return [Hash{Symbol => Object}]